fix const correctness of MeshTensor
